extract createCreateClient factory
